+++ src/channel_builder.rs
@@ -100,7 +100,7 @@ impl ChannelBuilder {
         //now compute the window based on our new bucket size
         let buckets_per_window:f32 = window.as_micros() as f32 / refresh_in_micros as f32;
         //find the next largest power of 2
-        let window_in_bits = (buckets_per_window as f32).log2().ceil() as u8;
+        let window_in_bits = (buckets_per_window).log2().ceil() as u8;
         result.with_compute_refresh_window_bucket_bits(refresh_in_bits, window_in_bits)
     }
 
@@ -113,38 +113,81 @@ impl ChannelBuilder {
         result
     }
 
+    /// Sets the capacity for the channel being built.
+    ///
+    /// # Parameters
+    /// - `capacity`: The maximum number of messages the channel can hold.
+    ///
+    /// # Returns
+    /// A `ChannelBuilder` instance with the specified capacity.
+    ///
+    /// # Notes
+    /// Increase capacity to reduce the chance of backpressure in high-throughput scenarios.
     pub fn with_capacity(&self, capacity: usize) -> Self {
         let mut result = self.clone();
         result.capacity = capacity;
         result
     }
 
-    /// show the type of T for this label
+    /// Enables type display in telemetry.
+    ///
+    /// # Returns
+    /// A `ChannelBuilder` instance with type display enabled.
+    ///
+    /// # Notes
+    /// Useful for debugging and telemetry analysis to identify the data types in channels.
     pub fn with_type(&self) -> Self {
         let mut result = self.clone();
         result.show_type = true;
         result
     }
 
-    /// expand the line weight as more data is processed over it
+    /// Enables line expansion in telemetry visualization.
+    ///
+    /// # Returns
+    /// A `ChannelBuilder` instance with line expansion enabled.
+    ///
+    /// # Notes
+    /// Line expansion visualizes the volume of data over time, enhancing traceability and insights.
     pub fn with_line_expansion(&self) -> Self {
         let mut result = self.clone();
         result.line_expansion = true;
         result
     }
 
+    /// Enables average calculation for the filled state of channels.
+    ///
+    /// # Returns
+    /// A `ChannelBuilder` instance with average filled calculation enabled.
+    ///
+    /// # Notes
+    /// Average calculations are lightweight and provide a basic understanding of channel utilization.
     pub fn with_avg_filled(&self) -> Self {
         let mut result = self.clone();
         result.avg_filled = true;
         result
     }
 
+    /// Enables average rate calculation.
+    ///
+    /// # Returns
+    /// A `ChannelBuilder` instance with average rate calculation enabled.
+    ///
+    /// # Notes
+    /// Useful for monitoring the average throughput of channels, identifying bottlenecks or underutilization.
     pub fn with_avg_rate(&self) -> Self {
         let mut result = self.clone();
         result.avg_rate = true;
         result
     }
 
+    /// Enables average latency calculation.
+    ///
+    /// # Returns
+    /// A `ChannelBuilder` instance with average latency calculation enabled.
+    ///
+    /// # Notes
+    /// Latency calculations help in assessing the responsiveness of the system and identifying delays.
     pub fn with_avg_latency(&self) -> Self {
         let mut result = self.clone();
         result.avg_latency = true;
@@ -152,6 +195,13 @@ impl ChannelBuilder {
     }
 
 
+    /// Marks this connection as going to a sidecar, for display purposes.
+    ///
+    /// # Returns
+    /// A `ChannelBuilder` instance configured with this mark
+    ///
+    /// # Notes
+    /// On charts and graphs, these two nodes will be in the same rank more 'near' each other than unrelated nodes.
 
     pub fn connects_sidecar(&self) -> Self {
         let mut result = self.clone();
@@ -159,59 +209,175 @@ impl ChannelBuilder {
         result
     }
 
-
+    /// Sets labels for the channel, optionally displaying them in telemetry.
+    ///
+    /// # Parameters
+    /// - `labels`: Static slice of label strings.
+    /// - `display`: Boolean indicating whether to display labels in telemetry.
+    ///
+    /// # Returns
+    /// A `ChannelBuilder` instance with specified labels.
+    ///
+    /// # Notes
+    /// Labels enhance telemetry data with descriptive tags, improving observability.
     pub fn with_labels(&self, labels: &'static [& 'static str], display: bool) -> Self {
         let mut result = self.clone();
         result.labels = if display {labels} else {&[]};
         result
     }
 
-
+    /// Configures the channel to calculate the standard deviation of the "filled" state.
+    ///
+    /// # Parameters
+    /// - `config`: Configuration for standard deviation calculation.
+    ///
+    /// # Returns
+    /// A modified instance of `ChannelBuilder` with filled state standard deviation calculation enabled.
+    ///
+    /// # Notes
+    /// Standard deviation for the filled state provides insights into the variability of the channel's usage over time.
+    /// It's a cost-effective way to gauge the consistency of channel capacity utilization without the computational overhead of histograms.
     pub fn with_filled_standard_deviation(&self, config: StdDev) -> Self {
         let mut result = self.clone();
         result.std_dev_filled.push(config);
         result
     }
+
+    /// Configures the channel to calculate the standard deviation of message rate.
+    ///
+    /// # Parameters
+    /// - `config`: Configuration for standard deviation calculation of message rate.
+    ///
+    /// # Returns
+    /// A modified instance of `ChannelBuilder` with message rate standard deviation calculation enabled.
+    ///
+    /// # Notes
+    /// This measures the variability in the rate at which messages are sent or received, offering insights into traffic patterns.
+    /// Ideal for identifying periods of high volatility in message throughput.
     pub fn with_rate_standard_deviation(&self, config: StdDev) -> Self {
         let mut result = self.clone();
         result.std_dev_rate.push(config);
         result
     }
+
+    /// Configures the channel to calculate the standard deviation of message latency.
+    ///
+    /// # Parameters
+    /// - `config`: Configuration for standard deviation calculation of latency.
+    ///
+    /// # Returns
+    /// A modified instance of `ChannelBuilder` with latency standard deviation calculation enabled.
+    ///
+    /// # Notes
+    /// Latency standard deviation helps identify fluctuations in message processing times, useful for performance tuning.
+    /// It provides a measure of how consistently the system processes messages, pinpointing instability in handling time.
     pub fn with_latency_standard_deviation(&self, config: StdDev) -> Self {
         let mut result = self.clone();
         result.std_dev_latency.push(config);
         result
     }
 
-
+    /// Configures the channel to calculate the standard deviation of message latency.
+    ///
+    /// # Parameters
+    /// - `config`: Configuration for standard deviation calculation of latency.
+    ///
+    /// # Returns
+    /// A modified instance of `ChannelBuilder` with latency standard deviation calculation enabled.
+    ///
+    /// # Notes
+    /// Latency standard deviation helps identify fluctuations in message processing times, useful for performance tuning.
+    /// It provides a measure of how consistently the system processes messages, pinpointing instability in handling time.
     pub fn with_filled_percentile(&self, config: Percentile) -> Self {
         let mut result = self.clone();
         result.percentiles_filled.push(config);
         result
     }
+
+    /// Configures the channel to calculate specific percentiles for message rate.
+    ///
+    /// # Parameters
+    /// - `config`: Configuration for the percentile calculation of message rate.
+    ///
+    /// # Returns
+    /// A modified instance of `ChannelBuilder` with rate percentile calculation enabled.
+    ///
+    /// # Notes
+    /// Rate percentiles provide a granular view of message throughput, highlighting the variability and extremes in message flow.
+    /// Essential for understanding peak and low traffic patterns, assisting in resource allocation and system scaling.
     pub fn with_rate_percentile(&self, config: Percentile) -> Self {
         let mut result = self.clone();
         result.percentiles_rate.push(config);
         result
     }
+
+    /// Configures the channel to calculate specific percentiles for message latency.
+    ///
+    /// # Parameters
+    /// - `config`: Configuration for the percentile calculation of latency.
+    ///
+    /// # Returns
+    /// A modified instance of `ChannelBuilder` with latency percentile calculation enabled.
+    ///
+    /// # Notes
+    /// Latency percentiles are key to understanding the distribution of message processing times, from best to worst.
+    /// They help in identifying and mitigating outliers that could affect user experience or system efficiency.
     pub fn with_latency_percentile(&self, config: Percentile) -> Self {
         let mut result = self.clone();
         result.percentiles_latency.push(config);
         result
     }
 
+    /// Configures triggers based on message rate with associated alert colors.
+    ///
+    /// # Parameters
+    /// - `bound`: The threshold for the trigger.
+    /// - `color`: The color to represent the alert when the threshold is crossed.
+    ///
+    /// # Returns
+    /// A modified instance of `ChannelBuilder` with a rate trigger configured.
+    ///
+    /// # Notes
+    /// Rate triggers alert when message throughput exceeds or falls below specified thresholds, enabling real-time reaction to traffic anomalies.
+    /// Ideal for systems requiring dynamic scaling or immediate notification of unusual activity.
     pub fn with_rate_trigger(&self, bound: Trigger<Rate>, color: AlertColor) -> Self {
         let mut result = self.clone();
         //need sequence number for these triggers?
         result.trigger_rate.push((bound, color));
         result
     }
+
+    /// Configures triggers based on the "filled" state of the channel with associated alert colors.
+    ///
+    /// # Parameters
+    /// - `bound`: The threshold for the trigger.
+    /// - `color`: The color to represent the alert when the threshold is crossed.
+    ///
+    /// # Returns
+    /// A modified instance of `ChannelBuilder` with a filled state trigger configured.
+    ///
+    /// # Notes
+    /// Filled state triggers provide immediate feedback on channel capacity issues, helping prevent backpressure and message loss.
+    /// They are crucial for maintaining optimal system performance and ensuring smooth message processing.
     pub fn with_filled_trigger(&self, bound: Trigger<Filled>, color: AlertColor) -> Self {
         let mut result = self.clone();
         //need sequence number for these triggers?
         result.trigger_filled.push((bound, color));
         result
     }
+
+    /// Configures triggers based on message latency with associated alert colors.
+    ///
+    /// # Parameters
+    /// - `bound`: The threshold for the trigger.
+    /// - `color`: The color to represent the alert when the threshold is crossed.
+    ///
+    /// # Returns
+    /// A modified instance of `ChannelBuilder` with a latency trigger configured.
+    ///
+    /// # Notes
+    /// Latency triggers are vital for systems where timely message processing is critical, allowing for swift identification and resolution of delays.
+    /// They help ensure that service level agreements (SLAs) are met and user satisfaction is maintained.
     pub fn with_latency_trigger(&self, bound: Trigger<Duration>, color: AlertColor) -> Self {
         let mut result = self.clone();
         //need sequence number for these triggers?
@@ -268,6 +434,9 @@ impl ChannelBuilder {
 
         }
     }
+
+    /// Finalizes the channel configuration and creates the channel with the specified settings.
+    /// This method ties together all the configured options, applying them to the newly created channel.
 
     pub fn build<T>(& self) -> (SteadyTx<T>, SteadyRx<T>) {
